Parameters to optionally correct small ENGINE DC hit handling errors.    {prefix}dc_fix_lr         Historically, in the ENGINE, if a hit is used in multiple         space points/stubs, the left/right assignment for that hit, which         is later used in track fitting, is that assigned for the last stub         encountered with that hit.  Properly, the left right assignment         should be allowed to be different in different space points.  If         this parameter is set to zero (e.g. in hcana.param), then the         ENGINE behavior is used.  For new analyses, it should be set to one.         (Which is the default)    {prefix}dc_fix_propcorr         When a hit gets put into a stub, the distance of the hit from the         discriminator can then be estimated.  In the engine, a correction to         the drift time (and thus drift distance) is applied.  However, if         that hit ends up in another stub, the correction will get applied again,         resulting in a over correction.  Setting this flag to 1 will give that         hit a different corection for each stub that it is in.
These flags will default to the new "correct" way of handling hits if
the above parameters are not set in a parameter file.  Currently, both
flags are set to zero in hcana.param to replicate the ENGINE behavior.

To implement these changes, the propagation correction and L/R information
for each hit is saved in space point and track classes.  This information
is still saved in the hit class, but only used if in ENGINE compatibility
mode.

The THcDCTrack class now saves a list of space point pointers instead
of space point indices.

The AddSpacePoint method now also copies all the hit information into
the track object so that THcDC doesn't need to explicitely copy all
the hits.

The FindStub method, which fits a stub track to a space point is passed
the space point rather than a list of hits