This is second update for THcCherenkov.cxx class. I have added four more variables. Total photo electrons, total hits of adc 1, total hits of adc 2 and total hits. With this update we have transformed the FORTRAN code of file /engine/HTRACKING/h_trans_cer.f to /hcana/ser/THcCherenkov.cxx
NOTE: The pedestal value in ENGINE is an integer but in hcana it is a
flaoting number. So it changes the pedestal substracted vlues of adc 1
and adc 2 and the number of photo electrons of adc 1 and 2 by very small
fraction. The difference of total photo electrons between ENGINE and hcana
is not more than 0.005.