Commiting changes in the raster branch before merging upstream.
Files with the prefix raster_ are coppies of original *.database, *.map and *.param files which now includes changes nessacery for decoding raster data. I created seperate files for these because I need to merge this code upstream before I can compare the raster decoding results with that of the analyzer. For e.g. the map files are different from the original map files. So Steve needs to create a perl script that can propgate these changes in the map files to the map file of the analyzer inorder for me to do the analyzer hcana comparision. Once the comparision is done and we are satisifies with the results, we will merge the changes with the original files and remove the files with the prefix raster_*.

Minor modifications in THcRaster* classes. Mostly changes done for debugging.