Golden track.   Goldent track is moved to THcHallCSpectrometer::TrackCalc()   if  fSelUsingScin == 0 then golden track is calculated just   like podd. i.e. it is the first track with minimum chi2/ndf   with sorting ON
  if fSelUsingScin == 1 then golden track is calculetd just like
  engine/HTRACKING/h_select_best_track_using_scin.h. This method
  gives the best track with minimum value of chi2/ndf but with
  additional cuts on the tracks. These cuts are on dedx, beta
  and on energy.

  This commit also corrects and modifies the beta code written
  in the THcHodoscope::FineProcess

  At this point we have godlend track and in additoion to
  previous vlues the following values are set to a track:
  Energy, beta, chi2 of beta and dedx[track][first hit].

NOTE: For some tracks we have large differences in value of chi
      square between ENGINE and hcana and this is already
      discussed in the last group meeting. Following emails are
      exchanged regarding this difference of chisquare.

email to Mark:

Hi Mark,
I tried to trace the origin of difference of the chi square value between
ENGINE and hcana. Please see the attached file, it has long table which I
cannot present in this email. Or you can see the file at the following
link, thanks.

https://userweb.jlab.org/~zahmed/chi2diff.txt

Ahmed

Mark's reply:
Ahmed,
        Thanks. I agree with what you see, but there is also another reason
for the difference. For the space point in ch2 the engine and hcana
choose different left/right combinations. That is why you see the large
jump in the summed chi2 from planes 9 and 10. The two codes pick
different left/right because there is a cut on the difference between the
measured stub xp and an expected xp = sp->GetX()/875  and one code has a combination
slightly below the cut value and the other slightly above. I do not see a way
to make the codes agree more exactly since the stub fitting involves a matrix
inversion.

                Cheers,
                       Mark

---------------------------------------------------------------------------------------