Look up of Shower and Cherenkov in THcHodoscope::Setup
Minimm value of normalized energy  fNormETot and minum value of number of
cherenkov photo electrons, fNCerNPE, is added to parameter files for the
calculation good scintillator event for tracking efficiency.
They are  optional parameters

fScinShould and fScinDid are set to 0 or 1
and only incrmented when Shower and Cerenkov detectors are setup.