Look up for shower and cherenkov in THcHodoscope::Setup - 2
If calorimeter and cherenkov is not added to SOS or HMS then it prints
a warning message and does not calcualte the efficiency for that
spectrometer.

Minimm value of normalized energy and minum value of number of
cherenkov photo electrons is added to parameter files for the
calculation of tracking efficiency.