Update SConstruct for evio handling and new version of podd
Podd no longer has the hana_scaler directory.

Also, on the advice of Stephen Wood, remove references to INSTALL_DIR (evio)
and now only check for EVIO_LIBDIR and EVIO_BINDIR.  Also, suppress
compilation of the evio libraries (as those directories may not be world
writeable.  So, the assumption now is that local installations of the
evio libraries will be done correctly (by the local user or the local
system administrator).