Added THcRawAdcHit and included it in THcRawShowerHit.
The new class THcRawAdcHit is meant to represent the raw ADC hits. The
logic with dealing with ADCs has become quite complex now and it would
have to be multiplicated across all detectors that have ADC
channels. While this by itself is not such a big problem, fixing
possible bugs in multiple places is.

I also redesigned the THcRawShowerHit class to use the new
THcRawAdcHit. The redesign was due anyway, so I took the opportunity.

I also had to modify THcShowerArray and THcShowerPlane. They still
need some work to use the the new flash 250 data, but changes were
necessary anyway.

Minor fix 01.