Replaced `gHcParms->Find()` calls with `gHcParms->LoadParmValues()`.
The calls to `Find()` need to be wrapped in checks if returned pointer is null.
Otherwise a further call to get value produces SegFault. The `LoadParmValues()`
does not suffer from this, since it throws error if non-optional parameter is
not found. The error message also tells which parameter is missing.

I still left three calls to `Find()` in `THcAnalyzer.cxx`. They are wrapped in
checks and some other things...