Fixed "No Time to dist algorithm available" error in `THcDCHit`.
This error is triggered when there is no wire defined for hit, or
when the wire has no `GetTDDConv()` method. The ROOT dictionary
calls default constructor that doesn't set the wire and calls
`ConvertTimeToDist()` regardless.

I changed the default constructor to not call the `ConvertTimeToDist()`
if the wire is not set. I also added the call to this function
in the `SetWire()` method.