Fixed a weird bug.
There was a weird bug in the code, that only triggered for some runs.
It turns out, in the `THcHodoscope::FineProcess` the hits in each plane
were created only for number of paddles in first plane. SHMS, however,
has 13 paddles in first plane and 14 and 21 paddles in last two planes.

The problem is fixed now, but part of that code should be rewritten for
efficiency.