Refactored fScinHitPaddle handling.
fScinHitPaddle is a std::vector<std::vector<Double_t>>. It was
cleared and rebuilt for each event with the same structure,
instead of just zeroing out all the values.

Now it is created in only one place, and than zeroed instead of
cleared.

Changed data type in fScinHitPaddle from Double_t to Int_t. Should
be changed to bool, but then all logic should be changed.