Added capability to set F250 parameters from PSE125.
PSE125 holds information about setting of F250 modules. Now this
information is used to set `fNPedestalSamples` and `fNPeakSamples`
parameters of `THcRawAdcHit`.

This is done through `THcHitList`, since it has information about
which crate and module each channel is in. Currently, the information
is set for each event, because of the hitlist and analyzer design.

I added the neccessary support for these changes to all raw hit
classes.