Complete rewrite of Aerogel and Cerenkov detector classes
New parameters are added.  hallc_replay of 3/31/2017 or later is needed.

As of now, the size of the TClonesArray objects are hard-coded in the
respective header files. This is not okay in the long run. Ideally,
the size of the TClonesArrays should be defined by the parameter which
dictates how many readout detectors exist for the respective
detector. I tried doing this in the Init method however, podd
complained regarding the size of the TClonesArrays in the DEF-file
histograms. I was unsuccessful in my first attempt. This should be
addressed in the future even though things work for now.