Add configuration option HCANA_BUILTIN_PODD (default ON)
If set to OFF (cmake -DHCANA_BUILTIN_PODD=OFF ...), CMake will find a
pre-built version of Podd according to the usual find_package search
rules. (It's sufficient to have the installed "analyzer" in your PATH,
but you can also set CMAKE_PREFIX_PATH to the Podd installation root,
for example.) This version of Podd has to have been built and
installed with Podd's CMake system. In this case, only libHallC will
be built, and the Podd submodule of hcana is ignored and does not even
have to be present.