Update THcHallCSpectrometer.cxx
Added setting of spectrometer mispointing according the the input
spectrometer angle if the mispointing is not set by a parameter.

The formulas for deteriming the mispointing comes from fits to sureys.

If one wants to use different spectrometer mispointings then
it is best to set hmispointing_x,hmispointing_y for HMS
and pmispointing_x,pmispointing_y for SHMS in the kinematics setting file
such as standard.kinematics .