Fix problems in THcTrigDet.cxx
1) Read-in of the trigger reference time cuts was in ReadDatabase method
  when it should be in Init method. Moved to the Init method before
  call to InitHitList where the reference cuts are used.

2) Modified the selection of the good TDC hit for filling the TdcTimeRaw
and TdcTime. Previosly used the TdcTime but not picks the TDC hit with the TdcTimeRaw that is within
the time min and max windows that can be set by parameters. Default windows are min=0 and max=10000.
If two hits in the time window then selects last hit within window.
If not hits within time window then TdcTimeRaw=0 and TdcTime=0.
The TdcMultiplicity is always filled.