Modify THcDC
In THcDC.h change MAXTRACKS from 10 to 50. The maximum
number of tracks is set by parameter ntracks_max_fp

Previously the LinkStubs method would not create
a track if the total number of spacepoints in
the two chambers was larger than 10.

Now LinkStubs will create a vector of THcSpacePoints
of up to 100 spacepoints with no more than
50 spacepoints from chamber 1 and use that
to find possible tracks.

I