Add THcScintPlaneClust, Modify THcHodoscope and THcScintillatorPlane
THcScintPlaneCluster is a new class to hold information
about scintillator plane clusters of paddle hits.
The clusters of paddles for each plane are formed in
THcHodoscope::TrackEffTest which is used in the tracking
efficiency.

Modify THcScintillatorPlane so that the cluster information
(number of clusters, size,pos, flag if cluster used)
is in the tree.

Modify THcHodoscope::TrackEffTest so position
difference  are taken for all clusters in X1 (or Y1)
with all clusters in X2 (Y2).