Updated THcHallCSpectrometer.cxx
In method TrackCalc get the proper THcDCTrack array index
from the golden track by using method GetTrkNum
for use in the THcDC method
SetFocalPlaneBestTrack.

When creating the THaTrack array in THcDC::CoarseTrack
the array index between THaTrack and THcDCTrack arrays
 match. The THaTrack is also associated with the THcDCTrack
array index with the call SetTrkNum(index+1)

When selecting best track by lowest chi-squared the
THaTrack are sorted from lowest to highest chi-squared
The code made a mistake in the index used in
 SetFocalPlaneBestTrack. It used the array index of the
THaTrack which is now not in sync with the THcDCTrack.
Fixed problem by using the THaTrack method GetTrkNum
which gets the proper THcDCTrack array index.


This did NOT effect reconstructed target variables or
physics variables. It only effected the focal plane
varaiables that were selected as the "best" or "golden"
to be used for the tree for dc.x_fp , dc.xp_fp,
dc.y_fp , dc.yp_fp and for the residuals.