Update ThcDriftChamber::LeftRight
The LeftRight method determines the sign of the wire
distance for each hit in the spacepoint by looping
through all combinations of wire signs and fitting each
combination with a line  and selecting
the combination with the lowest chi-squared.
The fit to the line is known as the "stub" which
is an array of the x,dx/dz, y and dy/dz of the fit
and is associated each spacepoint. The stubs are
used later in LinkStubs method
to match spacepoints between two chambers to give a
possible tracks.

Previously the LeftRight method would only fit a
stub for a spacepoint with 4 hits in the chamber if the
parameter flag  HMSStyleChamber was true.
HMSStyleChamber refers to the 6 GeV HMS chambers
and we do not have them anymore.
But not fitting the spacepoint, the stub
was filled with zeros for x,dx/dz,y and dy/dz
and these spacepoints would never be matched with
other spacepoints to form a possible track.