Modify THcHallCSpectrometer and THcDC
1) THcHallCspectrometer
a) add SetSHMSDipoleExitWindow which creates the TCutG
fDipoleExitWindowCutG based on SHMS exit window size
and sets fDipoleExitWindowZpos which is distance from
focal plane to SHMS exit window. Called in ReadDatabase.
b) add SetHMSDipoleExitWindow which creates the TCutG
fDipoleExitWindowCutG based on HMS exit window size
and sets fDipoleExitWindowZpos which is distance from
focal plane to HMS exit window. Called in ReadDatabase.
c) add Bool_t InsideDipoleExitWindow method which calculates the x and y
position at the exit window and checks if it is inside
the fDipoleExitWindowCutG
d) Add another check to the BestTrackUsingPrune method
to flag the tracks are inside the exit. Parameter
prune_DipoleExit needs to be set to 1 to use this check.
By default prune_DipoleExit is set to 0.

Modify THcDC.cxx
1) Added Bool_t tree variable dc.InSideDipoleExit
=  fInSideDipoleExit_best which is
set in method SetFocalPlaneBestTrack. This method calls
THcHallCSpectrometer::InsideDipoleExitWindow to check
whether the best track was inside the dipole
exit window and sets fInSideDipoleExit_best.