Modify THcDC and THcHallCSpectrometer
1) THcDC
a) Add variable to tree called dc.InsideDipoleExit
which is a Bool_t flag that is kTRUE if golden track is inside the
dipole exit. Uses THcHallCSpectrometer::InsideDipoleExitWindow
in THcDC::SetFocalPlaneBestTrack to set variable.

2) THcHallCSpectrometer
a) Add method Bool_t InsideDipoleExitWindow which checks whether
the track is within the dipole exit window. Checks
prefix used for database to set flag
fUseSHMSDipoleExitWindow or fUseHMSDipoleExitWindow
to be true which is used in InsideDipoleExitWindow
to select either SHMSDipoleExitWindow or
HMSDipoleExitWindow method to be used as the test.
b) Hardcoded z distance from spectrometer focal plane to
dipole exit window.
c) Methods Bool_t SHMSDipoleExitWindow and
Bool_t HMSDipoleExitWindow return true/false if
the track was inside/outside the dipole exit window.
d) Added dipole exit window test to the BestTrackUsingPrune
method. Need to set parameter hprune_DipoleExit=1 or
pprune_DipoleExit=1 to use the test in the method.
By default it is turned off.