Modified THcShower, THcShowerArray and THcShowerPlane
1) Add parameter fPedDefault which can be used as the
pedestal. When the raw pulse amplitude = 0, this means
that there was a pulse in the ADC time window where
the pedestal is determined. This makes the pedestal
wrong. so use the fPedDefault which is the average pedestal.

2) Use hodoscope OffsetTime in calculation of
the difference between the Hodoscope StartTime
and the ADC pulse time to compensate for difference
between ADC reference time and the TDC reference time.