Modify THCTrigDet::Decode
Previously if none of the TDC hits was not within the time cut
window the TDC value was set to 0. Now if there are
TDC hits it selections the first one in the cut window.
If no hit is found in the cut window then it selects the
hit that is closest in time to the lower window limit.