Update THcHodoscope fTrackBetaIncludeSinglePmtHits flag
When calculating beta using the track info the code
included hits from a paddle with single PMT hit.
Now the default behavior is to not to include these
single PMT hits in the track beta calculation.

The parameter p( or  h)TrackBetaIncludeSinglePmtHits = 1
will have the code ncluded hits from a paddle with single PMT hit
when calculating track beta.