Modify THcDriftChamberPlane.cxx and h
Add parameters p(h)min_drifttime (fMin_DriftTime)
and p(h)max drifttime (fMax_DriftTime) that puts a time cut
on selecting hits in SubtractStartTime method.
The default vaules are -50 and 200ns.

Now the code fills a TClonesArray fFirstPassHits
in ProcessHits with a time cut fTdcWinMin and fTdcWinMax
on the time that is not corrected for the starttime.
It use to fill fHits array.

In SubtractStartTime, the time is corrected for the hodoscope
starttime and if it passes the time cut fMin_DriftTime
and fMax_DriftTime, then fHits is filled.

fHits is used in the rest of the code to determine the track.