Fix problem with setting track momentum in THcHallCSpectrometer::FindVertices
Need to have the track momentum in the Hodoscope:CoarseProcess

Previously were setting track momentum in THcHallCSpectrometer::FindVertices
which is call AFTER Hodoscope:CoarseProcess.
By chance , only beta was calculated and since the track
momentum was kBig it cancelled to give beta = 1.

Moved calculation of track momentum to THcDC::CoarseTrack
which is called before Hodoscope:CoarseProcess