clean up THcRawAdcHit and THcCherenkov
Fix the cases when pulse is near the beginning or
end of the sample time window

In THcRawAdcHit::SetSampIntTimePedestalPeak
previouse always used fNPeakSamples = NSA+NSB
now set
 fNPeakSamples = TMath::Min((NS+fNSA-1),int(fNSamples-1)) - TMath::Max(NS-fNSB,0)+1;
in case the pulse is near the edge of the window.

In ThcCherenkov::Decode

For the case when FADC code returns Amp =0 for when the
pulse is near the far edge of the time window, set
the error flag ==2 when there is a waveform.
Previously it was set to 1.