Modify THcDC::DefineVariables
Remove track_hitpos as a tree variable.
Gave the warning.

Warning in <THaVarList::DefineVariables
[called from THcDC("H.dc")::DefineVarsFromList]>:
Too many dots in definition of variable
H.dc.track_hitpos (Position of each hit in track).
Variable not defined.

Introduced another way to get the hit positions for a track
but forgot to remove this variable.