Update for new FADC firmware installed in summer 2021 (#484)
* Update for FADC new firmware

* Updated RawADcHit

* Updated THcCherenkov

* Updated THcCherenkov so Errorflag = 3 for new failure mode

* Update raster code

* Updated THcRaster

* Updated THcRaster

* Update THcRawAdcHit, THcRawHit and THcRawHodoHit fro FADC mode 10

* Update THcHitList

Modify so the FADC reference time can be determined
from FADC if there is an FADC error which causes hit to
have  no pulse info and only waveform information
This is feature of the new FADC firmware.

* Update THcCherenkov to process FADC mode10 data

* Tweaked cointime module for PionLT run

* Added comment on coin time changes. Switched THcCherenkov to match origin version

* Tweaked naming of THcCoinTime variables to make it a little more readable - follows suggestion from Mark Jones. Switched TrigDet naming around, can set elsewhere

* Corrected a minor error in THcCoinTime, added RAW untracked variables since they weren't in this branch

* Modify THcCoinTime::Process

Fix last pTRIG4 to pHMS change

* Modify THcCherenkov::Decode

Fix mistake it filling the fSampWaveform

* Update FADC Waveform analysis

Modify THcRawAdcHit to have function SetSampNSAT
which sets the variable fNSAT. This is the Number of
consecutive Samples Above Threshold needs to find a pulse.
Default value is fNSAT = 2 in THcRawAdcHit::SetF250Params

Modify THcCherenkov
Add varaibles and parameters
fUseSampWaveform  det_UseSampWaveform 0= do not use , 1 = use Sample Waveform
fSampNSA          det_SampNSA default = 0 and use event 125 info
fSampNSAT         det_SampNSAT default = 2
fSampNSB          det_SampNSB default = 0 and use event 125 info

If det_UseSampWaveform = 1 then decode will fill the pulse data
structure with the data from the waveform which will be used
to determin the number of photoelectrons.

* clean up THcRawAdcHit and THcCherenkov

Fix the cases when pulse is near the beginning or
end of the sample time window

In THcRawAdcHit::SetSampIntTimePedestalPeak
previouse always used fNPeakSamples = NSA+NSB
now set
 fNPeakSamples = TMath::Min((NS+fNSA-1),int(fNSamples-1)) - TMath::Max(NS-fNSB,0)+1;
in case the pulse is near the edge of the window.

In ThcCherenkov::Decode

For the case when FADC code returns Amp =0 for when the
pulse is near the far edge of the time window, set
the error flag ==2 when there is a waveform.
Previously it was set to 1.

* Update THcAerogel class for sample data

Modified THcAerogel class to analyze the F250 Sample data.

* Updated THcShowerArray and THcShowerPlane to analyze F250 sample data.

* Update THcRawAdcHit::SetSampIntTimePedestalPeak

Add if statement so fSampThreshold ==0 then just returns the
pedestal and the integral is fixed to just above the pedestal
region. Used for determine pedestal needed for setting FADC
threshold. Data taken with all thresholds = 0.

* Updated THcTrigDet for mode 10 FADC data

* Updated THcScintillatorPlane to analyze F250 sample data.

* Fix typo in THcScintillatorPlane::ProcessHit

Co-authored-by: Cdaq User <cdaq@jlab.org>
Co-authored-by: David Hamilton <dhamilto@jlab.org>