Fix missing TI time
PipeliningModule::LoadBank sometimes supplies a short event length.
This happens because the TI time in the TI event sometimes looks like
a block trailer word.

Instead of looking at the supplied event length to determine if the event
contains the TI trigger time, we look at bit 16 of block header two which
indicates that the event contains the time.

There will still potentially be problem with multiblock data, but that would
have to be addressed in PipeliningModule or with a TI firmware change.