Get helicity information from raw sample data if there is no pulse data.
With a recent FADC firmware change, the FADC does not supply pulse data
for DC inputs such as the helicity signals.  Thus we can't get the helicity info
this way.  If pulse data is not seen, switch to a mode where the first raw
sample is used to get the pedestal.  When this is done, the fadc threshold
used to distinguish between 0 and 1 is divided by 4 as the pulse data used 4
samples to get the pedestal.

The user can force raw sample mode by setting helicity_adcrawsamples to 1 and
setting helicity_adcthreshold to 2000.