Helicity decoding fixes (#488)
* Fix missing TI time

PipeliningModule::LoadBank sometimes supplies a short event length.
This happens because the TI time in the TI event sometimes looks like
a block trailer word.

Instead of looking at the supplied event length to determine if the event
contains the TI trigger time, we look at bit 16 of block header two which
indicates that the event contains the time.

There will still potentially be problem with multiblock data, but that would
have to be addressed in PipeliningModule or with a TI firmware change.

* Get helicity information from raw sample data if there is no pulse data.

With a recent FADC firmware change, the FADC does not supply pulse data
for DC inputs such as the helicity signals.  Thus we can't get the helicity info
this way.  If pulse data is not seen, switch to a mode where the first raw
sample is used to get the pedestal.  When this is done, the fadc threshold
used to distinguish between 0 and 1 is divided by 4 as the pulse data used 4
samples to get the pedestal.

The user can force raw sample mode by setting helicity_adcrawsamples to 1 and
setting helicity_adcthreshold to 2000.