Modify ThcRawAdcHit:SetSampIntTimePedestalPeak()
1) change so pedestal is set to the smaller of
the first  fNPedestalSamples or the
last fNPedestalSamples in the sample window

2) If it finds a pulse, then the code looks for the
next pulse only after after it finds a sample
which is below threshold. This eliminates the
artifact of finding a 2nd or 3rd pulse from
the tails of a large pulse that are higher than
threshold past the NSA samples.