Add Event 137 handler (#520)
In the recent coda data, the new config event type (138) has been added. This event type includes configurations for readout modules that were used to be saved in event type 125. In particular, the parameters for FADC250 (such as NSA and NSB) are needed and looked up from decoding process in THcHitList.

* Add Evt137 handler to support the new FADC firmware
* Update THcHitList to add the new event 137 handler
* Add config output tree for event 137
* Update to use THaRunBase::GetDAQConfig